komet: migrate interval-role actions to BigDecimal (decimal-interval)
elk-snomed 0.40.0 + tinkar-core changed the interval API to BigDecimal bounds
(LogicalExpressionBuilder.IntervalRole/updateIntervalRoleValue, elk Interval
setters). Port the two komet/rules consumers that still passed int:
- AddIntervalRole: default bounds 0/10 -> BigDecimal.valueOf
- ChangeIntervalBound: parse user input as BigDecimal (decimal bounds);
  message 'Enter an integer' -> 'Enter a number'

Verified locally: komet rules compiles against elk 0.40.0.

Fixes: IKE-Network/ike-issues#688

+++ rules/src/main/java/dev/ikm/komet/rules/actions/axiom/AddIntervalRole.java
@@ -15,6 +15,8 @@
  */
 package dev.ikm.komet.rules.actions.axiom;
 
+import java.math.BigDecimal;
+
 import dev.ikm.komet.framework.panel.axiom.AxiomSubjectRecord;
 import dev.ikm.tinkar.entity.graph.adaptor.axiom.LogicalAxiom;
 import dev.ikm.tinkar.entity.graph.adaptor.axiom.LogicalExpressionBuilder;
@@ -39,8 +41,8 @@ public class AddIntervalRole extends AbstractAxiomAction {
 	@Override
 	public void doAction(ActionEvent t, AxiomSubjectRecord axiomSubjectRecord, EditCoordinateRecord editCoordinate) {
 		LogicalExpressionBuilder leb = new LogicalExpressionBuilder(axiomSubjectRecord.axiomTree());
-		LogicalAxiom.Atom.TypedAtom.IntervalRole role = leb.IntervalRole(TinkarTerm.UNMODELED_ROLE_CONCEPT, 0, true, 10,
-				true, TinkarTerm.UNMODELED_ROLE_CONCEPT);
+		LogicalAxiom.Atom.TypedAtom.IntervalRole role = leb.IntervalRole(TinkarTerm.UNMODELED_ROLE_CONCEPT,
+				BigDecimal.valueOf(0), true, BigDecimal.valueOf(10), true, TinkarTerm.UNMODELED_ROLE_CONCEPT);
 		leb.addToFirstAnd(axiomSubjectRecord.axiomIndex(), role);
 		putUpdatedLogicalExpression(editCoordinate, leb.build());
 	}
